dbdeep-DA/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
emoji==2.10.1
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.0
fsspec==2025.3.2
google-api-core==2.25.0rc1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-firestore==2.20.2
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.31.1
idna==3.6
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.45
langchain-huggingface==0.1.2
langchain-pinecone==0.2.6
langchain-tests==0.3.14
langchain-text-splitters==0.3.7
langdetect==1.0.9
langsmith==0.3.42
lxml==5.1.0
MarkupSafe==3.0.2
marshmallow==3.20.2
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
orjson==3.10.18
packaging==23.2
pandas==2.2.2
pillow==11.2.1
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic-settings==2.8.1
pydantic_core==2.23.4
pymys==0.2
PyMySQL==1.1.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.2
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.4.2
scipy==1.15.2
sentence-transformers==4.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.14.0
syrupy==4.9.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.66.2
transformers==4.49.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2025.2
unstructured==0.12.6
unstructured-client==0.18.0
urllib3==1.26.18
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.16.0
yarl==1.20.0
zstandard==0.23.0
